feat: 채널 목록, 포스트 목록 페이지 구분
