+++ src/pages/PostListPage.tsx
@@ -1,16 +1,15 @@
-import { useState } from 'react';
+import { useLocation } from 'react-router-dom';
 import { List, ListItem } from '~/components/domain/List';
-import { useGetChannels, useGetPosts } from '~/services';
-import { Channel, Post } from '~/types/model';
+import { useGetPosts } from '~/services';
+import { Post } from '~/types/model';
 
 const PostListPage = () => {
-  const [channelId, setChannelId] = useState('');
-  const { data: channelLsit } = useGetChannels();
-  const { data: postList } = useGetPosts({ channelId });
+  const location = useLocation();
+  const { data: postList } = useGetPosts({
+    channelId: location.pathname.split('/')[2]
+  });
 
-  const handleChangeChannel = (channelId: string) => {
-    setChannelId(channelId);
-  };
+  console.log(location.pathname.split('/')[2]);
 
   const getPostTitle = (postTitle: string) => {
     try {
@@ -24,32 +23,6 @@ const PostListPage = () => {
 
   return (
     <>
-      {/* 채널 목록페이지에서 클릭하면 id를 넘겨주면서 사용*/}
-      {/* 채널 목록페이지가 들어갈 자리*/}
-      {channelLsit && (
-        <>
-          <h1 className="text-2xl">채널 목록</h1>
-          <List>
-            {channelLsit.map((channelItem: Channel) => (
-              <ListItem
-                key={channelItem._id}
-                onClick={() => handleChangeChannel(channelItem._id)}
-                className="rounded-md border-2 p-2"
-              >
-                <label>채널 이름</label>
-                <p>{channelItem.name}</p>
-
-                <label>채널 설명</label>
-                <p>{channelItem.description}</p>
-
-                <label>생성 시간</label>
-                <p>{channelItem.createdAt}</p>
-              </ListItem>
-            ))}
-          </List>
-        </>
-      )}
-
       {postList && (
         <>
           <h1 className="text-2xl">채널에 해당하는 포스트 목록</h1>
