+++ src/pages/index.ts
@@ -7,3 +7,5 @@ export { FollowPage } from './FollowPage';
 export { default as ChatPage } from './ChatPage';
 export { default as ChatDetailPage } from './ChatDetailPage';
 export { NotFoundPage } from './NotFoundPage';
+export { default as ChannelListPage } from './ChannelListPage';
+export { default as PostListPage } from './PostListPage';
