+++ src/routes/constants.ts
@@ -2,7 +2,8 @@ export const PATH = {
   HOME: '/',
   SIGNUP: '/signup',
   POST: '/posts/:postId',
-  POSTS: '/posts',
+  CHANNELS: '/channels',
+  POSTS: '/channels/:channelId',
   POST_EDIT: '/post-edit',
   PROFILE: '/profile/:userId',
   FOLLOW: '/follow',
