+++ src/routes/router.tsx
@@ -3,17 +3,18 @@ import { PATH } from './constants';
 import Layout from './Layout';
 import LayoutWithFooter from './LayoutWithFooter';
 import {
+  ChannelListPage,
   ChatDetailPage,
   ChatPage,
   FollowPage,
   HomePage,
   NotFoundPage,
   PostEditPage,
+  PostListPage,
   PostPage,
   ProfilePage,
   SignupPage
 } from '~/pages';
-import PostListPage from '~/pages/PostListPage';
 
 const router = createBrowserRouter([
   {
@@ -31,6 +32,7 @@ const router = createBrowserRouter([
         ]
       },
       { path: PATH.SIGNUP, element: <SignupPage /> },
+      { path: PATH.CHANNELS, element: <ChannelListPage /> },
       { path: PATH.POST, element: <PostPage /> },
       { path: PATH.POSTS, element: <PostListPage /> },
       { path: PATH.POST_EDIT, element: <PostEditPage /> },
